requirements.txt
streamlit==1.54.0

torch==2.6.0
torchvision==0.21.0
ultralytics>=8.4.0

opencv-python-headless==4.9.0.80

numpy>=1.26,<2
pandas>=2.1,<3.0
pillow>=10.0
matplotlib>=3.8,<3.11
altair>=5,<6